Changed the files of getRewardsActivityTitle to a new file in utils
